like support for comments

+++ lib/notification-handlers/deleteNotificationOnUnlike.ts
@@ -2,12 +2,25 @@ import { PrismaClient } from "@prisma/client";
 
 const prisma = new PrismaClient();
 
-export default async (sender: string, postId: string) => {
-  await prisma.notification.deleteMany({
-    where: {
-      sender,
-      postId,
-      type: "LIKE"
-    }
-  })
+export default async (sender: string, postId: string|null, commentId: string|null) => {
+
+  if (postId) {
+    await prisma.notification.deleteMany({
+      where: {
+        sender,
+        postId,
+        type: "LIKE",
+      },
+    });
+  }
+
+  if (commentId) {
+    await prisma.notification.deleteMany({
+      where: {
+        sender,
+        commentId,
+        type: "LIKE",
+      },
+    });
+  }
 };
